+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Changelog
 
 **Next release**
 
+* Toggle restriction will allow to choose between, clear, restrict and template ([issue](/../../issues/1290))
+* Updated Dutch translation
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.53 TEST**